+++ examples/resources/packetfabric_cloud_router_connection_aws/resource.tf
@@ -1,16 +1,14 @@
 resource "packetfabric_cloud_router" "cr1" {
-  provider     = packetfabric
-  asn          = var.pf_cr_asn
-  name         = var.pf_cr_name
-  account_uuid = var.pf_account_uuid
-  capacity     = var.pf_cr_capacity
-  regions      = var.pf_cr_regions
+  provider = packetfabric
+  asn      = var.pf_cr_asn
+  name     = var.pf_cr_name
+  capacity = var.pf_cr_capacity
+  regions  = var.pf_cr_regions
 }
 
 resource "packetfabric_cloud_router_connection_aws" "crc1" {
   provider       = packetfabric
   circuit_id     = packetfabric_cloud_router.cr1.id
-  account_uuid   = var.pf_account_uuid
   aws_account_id = var.pf_aws_account_id
   maybe_nat      = var.pf_crc_maybe_nat
   description    = var.pf_crc_description