+++ examples/use-cases/backbone_virtual_circuit_mesh/main.tf
@@ -2,15 +2,12 @@ terraform {
   required_providers {
     packetfabric = {
       source  = "PacketFabric/packetfabric"
-      version = ">= 0.4.0"
+      version = ">= 0.4.2"
     }
   }
 }
 
-provider "packetfabric" {
-  host  = var.pf_api_server
-  token = var.pf_api_key
-}
+provider "packetfabric" {}
 
 # Create random name to use to name objects
 resource "random_pet" "name" {}
@@ -31,7 +28,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc1" {
     vlan            = var.pf_vc_vlan1
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -54,7 +50,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc2" {
     vlan            = var.pf_vc_vlan2
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -80,7 +75,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc3" {
     vlan            = var.pf_vc_vlan3
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -103,7 +97,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc4" {
     vlan            = var.pf_vc_vlan4
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -126,7 +119,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc5" {
     vlan            = var.pf_vc_vlan5
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -149,7 +141,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc6" {
     vlan            = var.pf_vc_vlan6
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -172,7 +163,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc7" {
     vlan            = var.pf_vc_vlan7
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -195,7 +185,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc8" {
     vlan            = var.pf_vc_vlan8
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -218,7 +207,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc9" {
     vlan            = var.pf_vc_vlan9
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -241,7 +229,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc10" {
     vlan            = var.pf_vc_vlan10
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -264,7 +251,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc11" {
     vlan            = var.pf_vc_vlan11
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -287,7 +273,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc12" {
     vlan            = var.pf_vc_vlan12
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -310,7 +295,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc13" {
     vlan            = var.pf_vc_vlan13
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -333,7 +317,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc14" {
     vlan            = var.pf_vc_vlan14
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm
@@ -356,7 +339,6 @@ resource "packetfabric_backbone_virtual_circuit" "vc15" {
     vlan            = var.pf_vc_vlan15
   }
   bandwidth {
-    account_uuid      = var.pf_account_uuid
     longhaul_type     = var.pf_vc_longhaul_type
     speed             = var.pf_vc_speed
     subscription_term = var.pf_vc_subterm