+++ examples/use-cases/cloud_router_aws_google/cloud_router_connection_google.tf
@@ -30,7 +30,6 @@ resource "packetfabric_cloud_router_connection_google" "crc_2" {
   provider                    = packetfabric
   description                 = "${var.tag_name}-${random_pet.name.id}-${var.pf_crc_pop2}"
   circuit_id                  = packetfabric_cloud_router.cr.id
-  account_uuid                = var.pf_account_uuid
   google_pairing_key          = google_compute_interconnect_attachment.google_interconnect_1.pairing_key
   google_vlan_attachment_name = google_compute_interconnect_attachment.google_interconnect_1.name
   pop                         = var.pf_crc_pop2