+++ internal/provider/resource_azure_hosted_mkt_conn.go
@@ -54,8 +54,10 @@ func resourceAzureHostedMktConn() *schema.Resource {
 			"account_uuid": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: validation.StringIsNotEmpty,
-				Description:  "The UUID for the billing account that should be billed. This is your billing account, not the marketplace provider's.",
+				DefaultFunc:  schema.EnvDefaultFunc("PF_ACCOUNT_ID", nil),
+				ValidateFunc: validation.IsUUID,
+				Description: "The UUID for the billing account that should be billed. " +
+					"Can also be set with the PF_ACCOUNT_ID environment variable.",
 			},
 			"speed": {
 				Type:         schema.TypeString,