+++ internal/provider/resource_azure_req_express_conn.go
@@ -36,9 +36,12 @@ func resourceAzureReqExpressConn() *schema.Resource {
 			"account_uuid": {
 				Type:         schema.TypeString,
 				Required:     true,
+				DefaultFunc:  schema.EnvDefaultFunc("PF_ACCOUNT_ID", nil),
 				ValidateFunc: validation.IsUUID,
-				Description:  "The UUID for the billing account that should be billed.",
+				Description: "The UUID for the billing account that should be billed. " +
+					"Can also be set with the PF_ACCOUNT_ID environment variable.",
 			},
+
 			"description": {
 				Type:         schema.TypeString,
 				Required:     true,